First commit:
Created the speedLimit () function.
Assigned variables of speed and demerits. 
Created an if statement that defines 0km as the min and 70km as the max input that would print "OK".
Created an if statement assigning demerit points for every 5km over the speed limit.
Else {} to print out an error message.
Assigned a math.floor function to the demerits variable.