Second commit.
Second code challenge.
Created a Java Script function grade(). Created a variable for score an assigned the html input value to it, converting it into an integer using parseInt. Created an if statement to compartmentalize the scores and assign grades to each category. Printed the outcome into html using getElementById.