Third Commit: Third Code Challenge.
Created netSalaryCalculator function. Inside the function, I initially assigned two variables respectively to the inputs of basic salary and benefits. Used parseInt to convert to integer and .value to collect the accurate number to a radix base 10. Created the grossSalary variable that contains a sum of both basic salary and benefits. Created a paye variable is subsequent if statements for the PAYE deductions. I did the same for both NHIF and NSSF statements. Created netDeductions for all deductibles and personalTaxRelief to accommodate the tax relief value. Created netSalary variable for the resulting computations, printing string to html using getElementById().textContent.